# staccato-main/src/main/java/com/boundlessgeo/staccato/transaction/TransactionApi.java
package com.boundlessgeo.staccato.transaction;

import com.boundlessgeo.staccato.dto.StacTransactionResponse;
import com.boundlessgeo.staccato.model.Item;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.*;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;

/**
 *  Defines the controller interface for the STAC API specification
 *  @see <a href="https://github.com/radiantearth/stac-spec/tree/master/extensions/transaction">transaction extension</a>
 * @author joshfix
 * Created on 08/15/2018
 */
public interface TransactionApi {

    @PostMapping(value = "/collections/{collectionId}/items", consumes = MediaType.APPLICATION_JSON_VALUE)
    Mono<StacTransactionResponse> createItem(@RequestBody Flux<Item> items, @PathVariable("collectionId") String collectionId);

    @PutMapping(value = "/collections/{collectionId}/items/{itemId}", consumes = MediaType.APPLICATION_JSON_VALUE)
    Mono<StacTransactionResponse> createItem(@ModelAttribute Item item, @PathVariable("collectionId") String collectionId,
                                             @PathVariable("itemId") String itemId);

    @PatchMapping("/collections/{collectionId}/items/{itemId}")
    Mono<StacTransactionResponse> updateItem(@RequestBody String body, @PathVariable("collectionId") String collectionId,
                                             @PathVariable("itemId") String itemId);

    @DeleteMapping("/collections/{collectionId}/items/{itemId}")
    Mono<StacTransactionResponse> deleteItem(@PathVariable("collectionId") String collectionId, @PathVariable("id") String id);

}