# staccato-main/src/main/resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "An API for SpatioTemporal Asset Catalogs",
    "version": "1.0.0-RC10",
    "description": "The SpatioTemporal Asset Catalog specification aims to standardize the way geospatial assets are exposed online and queried.  A 'spatiotemporal asset' is any file that represents information about the earth captured in a certain space and time. Most of the initial  use cases are related to imagery, but the api is designed to enable search of any spatiotemporal data items. SpatioTemporal Asset Catalogs  can be completely static, implemented as flat files on the web that can be crawled. This API has the same responses as a static catalog, but also offers querying capabilities in addition to exposing data online. This API was inspired by WFS 3.0, and the goal is to make it compatible with that specification.",
    "contact": {
      "name": "Boundless Geo",
      "email": "jfix@boundlessgeo.com",
      "url": "https://github.com/radiantearth/stac-spec/"
    },
    "license": {
      "name": "CC-BY 4.0 license",
      "url": "https://creativecommons.org/licenses/by/4.0/"
    }
  },
  "servers": [
    {
      "description": "SwaggerHub API Auto Mocking",
      "url": "https://virtserver.swaggerhub.com/boundlessgeo/stac/1.0.0-RC10"
    },
    {
      "description": "SwaggerHub API Auto Mocking",
      "url": "http://stac.boundlessgeo.io/stac"
    }
  ],
  "paths": {
...
    "/items": {
      "get": {
        "summary": "Retrieve STAC compliant Items matching filters",
        "description": "Returns a pageable GeoJSON feature collection of the Items matching the query",
        "operationId": "getItems",
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/bbox"
          },
          {
            "$ref": "#/components/parameters/time"
          },
          {
            "$ref": "#/components/parameters/filter"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/next"
          },
          {
            "$ref": "#/components/parameters/propertyname"
          },
          {
            "$ref": "#/components/parameters/AcceptHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Description here",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemCollection"
                }
              }
            }
          },
          "default": {
            "description": "An error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Accepts a JSON representation of the query parameters to seach and produce an ItemCollection",
        "operationId": "getItemsPost",
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/requestBodies/JsonPostBody"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "string"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a pageable GeoJSON feature collection of the items matching the query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemCollection"
                }
              }
            }
          },
          "default": {
            "description": "An error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Adds one or more items to the catalog",
        "description": "The request body may contain either a single item in JSON format or an array of items.",
        "operationId": "addItem",
        "tags": [
          "Items"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a JSON response with the success status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "summary": "Retreive item by id",
        "operationId": "getItem",
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "An item.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "description": "The item with id {id} does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "default": {
            "description": "An error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete item by id",
        "operationId": "deleteItem",
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful delete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          },
          "404": {
            "description": "The item with id {id} does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "default": {
            "description": "An error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Allows partial updates to items",
        "operationId": "updateItem",
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/requestBodies/JsonPostBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful delete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          },
          "404": {
            "description": "The item with id {id} does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "default": {
            "description": "An error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "components": {
    "parameters": {
      "next": {
        "name": "next",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Token for the next set of paginated data."
      },
      "limit": {
        "name": "limit",
        "in": "query",
        "description": "Defines the maximum number or results retruned by the API",
        "schema": {
          "type": "integer"
        }
      },
      "time": {
        "name": "time",
        "description": "An array of date-time strings that adheres to RFC3339 (for example 1985-04-12T23:20:50.52Z). One date-time in the array is the time to match. Two date-times represent a range. Any additional items in the array will be ignored.",
        "in": "query",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "bbox": {
        "name": "bbox",
        "in": "query",
        "description": "Only items that have a geometry that intersects the bounding box are selected. The bounding box is provided as four numbers, in longitude & latitude:\n* West, South, East, North",
        "schema": {
          "type": "array",
          "items": {
            "type": "number",
            "minItems": 4,
            "maxItems": 4,
            "minimum": -180,
            "maximum": 180
          }
        }
      },
      "filter": {
        "name": "filter",
        "in": "query",
        "description": "A CQL string based off of GeoTools CQL language",
        "schema": {
          "type": "string"
        }
      },
      "propertyname": {
        "name": "propertyname",
        "in": "query",
        "description": "A comma separated list of property names to be returned in the JSON response",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "id": {
        "name": "id",
        "in": "path",
        "description": "The id of an item",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "AcceptHeader": {
        "in": "header",
        "name": "Accept",
        "description": "Setting the Accept header value to 'text/event-stream' enables Server-Sent Events. Instead of returning a single ItemCollection, each Item will be streamed back from the server as they become available.",
        "schema": {
          "type": "string"
        },
        "required": false
      }
    },
    "requestBodies": {
      "JsonPostBody": {
        "description": "JSON object containing query parameters",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/JsonPostRequest"
            }
          }
        }
      }
    },
    "schemas": {
      "JsonPostRequest": {
        "type": "object",
        "properties": {
          "bbox": {
            "description": "Array of doubles in order of minx, miny, maxx, maxy",
            "type": "array"
          },
          "limit": {
            "description": "An integer limiting the number of records to be returned.",
            "type": "integer"
          },
          "time": {
            "description": "A string array of date timestamps used to filter the returned records between two dates.",
            "type": "array"
          },
          "filter": {
            "description": "A CQL string used to filter the results",
            "type": "string"
          },
          "next": {
            "description": "An integer that can be used to get the next page of results. Can be used in conjuction with limit to provide for pagination of results.",
            "type": "integer"
          },
          "propertyname": {
            "description": "A string array of property names that should be returned in the response",
            "type": "array"
          }
        }
      },
      "exception": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "HttpResponse": {
        "description": "Response object for PUT and DELETE operations",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "result": {
            "type": "object"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "bbox": {
        "description": "West, South, East, North",
        "type": "array",
        "items": {
          "minItems": 4,
          "maxItems": 4,
          "type": "number",
          "minimum": -180,
          "maximum": 180
        },
        "example": [
          -180,
          -90,
          180,
          90
        ]
      },
...
      "ItemCollection": {
        "type": "object",
        "required": [
          "items",
          "type"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "ItemCollection"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "nextPageToken": {
            "type": "string",
            "description": "A token to obtain the next paginated data set",
            "example": "ANsXtp9mrqN0yrKWhf-y2PUpHRLQb1GT-mtxNcXou8TwkXhi1Jbk"
          },
          "links": {
            "$ref": "#/components/schemas/ItemCollectionLinks"
          }
        }
      },
      "Item": {
        "type": "object",
        "required": [
          "id",
          "type",
          "geometry",
          "links",
          "properties"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "path/to/example.tif",
            "description": "unique ID, potentially link to file"
          },
          "bbox": {
            "$ref": "#/components/schemas/bbox"
          },
          "geometry": {
            "type": "object",
            "description": "footprint of the item"
          },
          "type": {
            "type": "string",
            "description": "The GeoJSON type",
            "enum": [
              "Item"
            ]
          },
          "properties": {
            "$ref": "#/components/schemas/ItemProperties"
          },
          "links": {
            "$ref": "#/components/schemas/ItemLinks"
          }
        },
        "example": {
          "type": "Feature",
          "id": "CS3-20160503_132130_04",
          "bbox": [
            -122.59750209,
            37.48803556,
            -122.2880486,
            37.613537207
          ],
          "geometry": {
            "type": "Polygon",
            "coordinates": [
              [
                [
                  -122.308150179,
                  37.488035566
                ],
                [
                  -122.597502109,
                  37.538869539
                ],
                [
                  -122.576687533,
                  37.613537207
                ],
                [
                  -122.2880486,
                  37.562818007
                ],
                [
                  -122.308150179,
                  37.488035566
                ]
              ]
            ]
          },
          "properties": {
            "start": "2016-05-03T13:21:30.040Z",
            "end": "2016-05-03T13:21:30.410Z",
            "provider": "http://www.cool-sat.com",
            "license": "CC-BY-4.0"
          },
          "links": [
            {
              "rel": "self",
              "href": "http://cool-sat.com/catalog/CS3-20160503_132130_04/CS3-20160503_132130_04.json"
            },
            {
              "rel": "thumbnail",
              "href": "http://cool-sat.com/catalog/CS3-20160503_132130_04/thumbnail.png"
            }
          ],
          "assets": [
            {
              "href": "relative-path/to/analytic.tif",
              "name": "4-Band Analytic"
            }
          ]
        }
      },
      "ItemProperties": {
        "type": "object",
        "required": [
          "start_date",
          "end_date"
        ],
        "description": "provides the core metatdata fields plus extensions",
        "properties": {
          "start_date": {
            "$ref": "#/components/schemas/Time"
          },
          "end_date": {
            "$ref": "#/components/schemas/Time"
          },
          "provider": {
            "type": "string",
            "description": "Information about the provider"
          },
          "license": {
            "type": "string",
            "description": "Data license name based on SPDX License List"
          }
        },
        "additionalProperties": {
          "type": "object"
        }
      },
      "ItemLinks": {
        "type": "object",
        "required": [
          "self",
          "thumbnail"
        ],
        "properties": {
          "self": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          }
        }
      },
      "ItemCollectionLinks": {
        "type": "object",
        "properties": {
          "next": {
            "type": "string",
            "format": "url",
            "description": "A URL to obtain the next paginated data set. If not present, the client should utilize the `nextPageToken`.",
            "example": "http://api.cool-sat.com/query/gasd312fsaeg/ANsXtp9mrqN0yrKWhf-y2PUpHRLQb1GT-mtxNcXou8TwkXhi1Jbk"
          }
        }
      },
      "Time": {
        "type": "string",
        "description": "A single date-time string that adheres to RFC3339, for example 1985-04-12T23:20:50.52Z"
      },
...
    }
  }
}